TTCI/Classificação_KNN_API/API/requirements.txt
# Python 3.9.6
annotated-types==0.7.0
anyio==4.4.0
art==6.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.1
fastapi==0.108.0
h11==0.14.0
idna==3.7
joblib==1.4.2
numpy==1.26.3
opencv-python-headless==4.8.0.76
Pillow==9.4.0
pydantic==2.8.2
pydantic-core==2.20.1
python-multipart==0.0.7
scikit-learn==1.2.2
scipy==1.13.1
sniffio==1.3.1
starlette==0.32.0.post1
threadpoolctl==3.5.0
typing-extensions==4.12.2
uvicorn==0.25.0